requirements.txt
dlib-19.22.99-cp310-cp310-win_amd64.whl
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
face-recognition==1.3.0
pandas==2.2.2
numpy==1.24.3
Pillow==10.4.0
mediapipe==0.10.14
